refactor: PR 피드백 반영 - car.name 의 wrapper 클래스 생성. validation 이동 - main 함수 (어플리케이션 실행)클래스 분리 - 동일 위치에 대한 확인 Car 내부에서 수행 - 이름 validation 걸리는 단위테스트 및 기본값 수정
