+++ src/main/java/com/nextstep/javaRacing/racing/car/RandomMoveStrategy.java
@@ -7,7 +7,7 @@ public class RandomMoveStrategy implements MoveStrategy {
     private static final int MAX_RANDOM = 10;
     private static final int THRESHOLD = 4;
 
-    private final Random random = new Random();
+    private final Random random = new Random(System.currentTimeMillis());
 
     @Override
     public boolean isMovable() {
